chore(core): abstract tinybench
