+++ src/bootstrap.ts
@@ -6,7 +6,8 @@
  */
 
 import { ModestBenchConfigurationManager } from './config/manager.js';
-import { ModestBenchEngine } from './core/engine.js';
+import { type ModestBenchEngine } from './core/engine.js';
+import { TinybenchEngine } from './core/engines/index.js';
 import { ModestBenchErrorManager } from './core/error-manager.js';
 import { BenchmarkFileLoader } from './core/loader.js';
 import { ModestBenchProgressManager } from './progress/manager.js';
@@ -16,11 +17,13 @@ import { FileHistoryStorage } from './storage/history.js';
 /**
  * Initializes the ModestBench engine with default dependencies.
  *
+ * Uses TinybenchEngine as the default concrete implementation.
+ *
  * @returns {ModestBenchEngine} The initialized ModestBench engine.
  * @public
  */
 export const bootstrap = (): ModestBenchEngine => {
-  const engine = new ModestBenchEngine({
+  const engine = new TinybenchEngine({
     configManager: new ModestBenchConfigurationManager(),
     errorManager: new ModestBenchErrorManager(),
     fileLoader: new BenchmarkFileLoader(),
