+++ src/core/engines/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Benchmark Engine Implementations
+ *
+ * Concrete implementations of the ModestBenchEngine abstract class using
+ * different underlying benchmark libraries.
+ *
+ * @packageDocumentation
+ */
+
+export { TinybenchEngine } from './tinybench-engine.js';
